implemented jwt token for register and login
