+++ src/main/java/com/vietle/angularecommercebackend/controller/ItemController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.UUID;
 
 @RestController
@@ -25,8 +26,8 @@ public class ItemController {
     private ItemService itemService;
 
     @GetMapping("/item-detail/{id}/{category}")
-    public ResponseEntity<ItemResponse> retrieveItemDetial(@PathVariable int id, @PathVariable int category) throws EcommerceException {
-        Item item = this.itemService.retrieveItem(id, category);
+    public ResponseEntity<ItemResponse> retrieveItemDetial(@PathVariable int id, @PathVariable int category, HttpServletRequest req) throws EcommerceException {
+        Item item = this.itemService.retrieveItem(id, category, req);
         String transactionId = UUID.randomUUID().toString();
         Status status = Status.builder().statusCd(200).message(Constant.SUCCESS).transactionId(transactionId).timestamp(EcommerceUtil.getTimestamp()).build();
         ItemResponse itemResponse = ItemResponse.builder().item(item).status(status).build();
