+++ src/main/java/com/vietle/angularecommercebackend/repo/UserRepositoryImpl.java
@@ -1,27 +1,56 @@
 package com.vietle.angularecommercebackend.repo;
 
+import com.vietle.angularecommercebackend.domain.Role;
 import com.vietle.angularecommercebackend.domain.User;
 import com.vietle.angularecommercebackend.exception.EcommerceException;
 import com.vietle.angularecommercebackend.security.JwtTokenProvider;
 import com.vietle.angularecommercebackend.service.MongoSequenceService;
 import com.vietle.angularecommercebackend.util.EcommerceUtil;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Repository;
 
+import javax.annotation.PostConstruct;
+import java.util.Base64;
+import java.util.Date;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Repository
 public class UserRepositoryImpl implements UserRepository {
+    @Value("${security.jwt.token.secret-key}")
+    private String secretKey;
+    @Value("${security.jwt.token.expire-length}")
+    private long validityInMilliseconds;
     @Autowired
     private MongoTemplate mongoTemplate;
     @Autowired
     private MongoSequenceService mongoSequenceService;
 
-//    @Autowired
-//    private JwtTokenProvider jwtTokenProvider;
+
+    public String createToken(String username, List<Role> roles) {
+
+        Claims claims = Jwts.claims().setSubject(username);
+        claims.put("auth", roles.stream().map(s -> new SimpleGrantedAuthority(s.getAuthority())).filter(Objects::nonNull).collect(Collectors.toList()));
+
+        Date now = new Date();
+        Date validity = new Date(now.getTime() + validityInMilliseconds);
+
+        return Jwts.builder()//
+                .setClaims(claims)//
+                .setIssuedAt(now)//
+                .setExpiration(validity)//
+                .signWith(SignatureAlgorithm.HS256, secretKey)//
+                .compact();
+    }
 
     @Override
     public User save(User user) throws EcommerceException {
@@ -33,7 +62,8 @@ public class UserRepositoryImpl implements UserRepository {
         //TODO: check savedUser and handle exception
         savedUser.setPassword(null);
         //token
-//        String token = jwtTokenProvider.createToken(savedUser.getEmail(), user.getRoles());
+        String token = this.createToken(savedUser.getEmail(), user.getRoles());
+        System.out.println("GENERATED TOKEN: " + token);
         //
         return savedUser;
     }
@@ -70,6 +100,10 @@ public class UserRepositoryImpl implements UserRepository {
         List<User> foundUsers = this.mongoTemplate.find(query, User.class);
         User foundUser = foundUsers.stream().findFirst().orElseThrow(() -> new EcommerceException("invalid login!", 400));
         foundUser.setPassword(null);
+        //token
+        String token = this.createToken(foundUser.getEmail(), foundUser.getRoles());
+        System.out.println("GENERATED TOKEN: " + token);
+        foundUser.setToken(token);
         return foundUser;
     }
 }
