+++ src/main/java/com/vietle/angularecommercebackend/security/JwtTokenFilter.java
@@ -28,7 +28,6 @@ public class JwtTokenFilter extends OncePerRequestFilter {
                 SecurityContextHolder.getContext().setAuthentication(auth);
             }
         } catch (EcommerceException ex) {
-            //this is very important, since it guarantees the user is not authenticated at all
             SecurityContextHolder.clearContext();
             httpServletResponse.sendError(500, ex.getMessage());
             return;
