+++ src/main/resources/application.properties
@@ -7,8 +7,6 @@ server.servlet.context-path=/ecommerce
 logging.level.com.vietle.angularecommercebackend=INFO
 
 #jwt config
-jwt.token.secret=secret
-jwt.token.expire=300000 # 5 minutes duration by default: 5 minutes * 60 seconds * 1000 miliseconds
-
 security.jwt.token.secret-key=secret
+#valid for 5 minutes
 security.jwt.token.expire-length=300000
